added home styles to seperate file
