+++ Aura_frontend/app/(auth)/SignIn.tsx
@@ -54,7 +54,7 @@ const SignIn: React.FC = () => {
       console.log("Tokennn:", token);
 
       alert("Signed in successfully");
-      router.replace("/(tabs)/Home");
+      router.replace("/(tabs)/Home/Home");
     } catch (error: any) {
       const message =
         error?.response?.data?.message || "Sign in failed. Try again.";
