+++ Aura_frontend/app/(auth)/SignUp.tsx
@@ -52,7 +52,7 @@ const SignUp: React.FC = () => {
       await login(res.data.token);
       alert("Signed up successfully");
 
-      router.replace("/(tabs)/Home");
+      router.replace("/(tabs)/Home/Home");
     } catch (error: any) {
       const message =
         error?.response?.data?.message || "Sign up failed. Try again.";
